moving isValidUrl
