+++ packages/suite-base/src/util/isValidURL.ts
@@ -1,0 +1,14 @@
+// SPDX-FileCopyrightText: Copyright (C) 2023-2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)<lichtblick@bmwgroup.com>
+// SPDX-License-Identifier: MPL-2.0
+
+const VALID_PROTOCOLS = ["https:", "http:", "file:", "data:", "package:"];
+
+export function isValidUrl(str: string): boolean {
+  try {
+    const url = new URL(str);
+    return VALID_PROTOCOLS.includes(url.protocol);
+  } catch (err: unknown) {
+    console.error(err);
+    return false;
+  }
+}
